fix: Add unit test for ProgressCard
